Fix unnecessary casts.

+++ bundles/org.eclipse.swt/Eclipse SWT PI/win32/org/eclipse/swt/internal/win32/FLICK_DATA.java
@@ -3,8 +3,7 @@ package org.eclipse.swt.internal.win32;
 
 public class FLICK_DATA {
 	public int iFlickActionCommandCode;
-	/** @field cast=(unsigned int) */
-	public int iFlickDirection;
+	public byte iFlickDirection;
 	public boolean fControlModifier;
 	public boolean fMenuModifier;
 	public boolean fAltGRModifier;
